Move decision for CompositeCheckout to CheckoutSystemDecider
This will allow us to more easily pull data for the decision process.
