+++ client/my-sites/checkout/controller.jsx
@@ -15,16 +15,13 @@ import { getSiteBySlug } from 'state/sites/selectors';
 import { getSelectedSite } from 'state/ui/selectors';
 import GSuiteNudge from './gsuite-nudge';
 import CheckoutContainer from './checkout/checkout-container';
+import CheckoutSystemDecider from './checkout/checkout-system-decider';
 import CheckoutPendingComponent from './checkout-thank-you/pending';
 import CheckoutThankYouComponent from './checkout-thank-you';
 import UpsellNudge from './upsell-nudge';
 import { isGSuiteRestricted } from 'lib/gsuite';
 import { getRememberedCoupon } from 'lib/cart/actions';
 import { sites } from 'my-sites/controller';
-import config from 'config';
-import CompositeCheckout from './checkout/composite-checkout';
-import CartData from 'components/data/cart';
-import { abtest } from 'lib/abtest';
 
 export function checkout( context, next ) {
 	const { feature, plan, domainOrProduct, purchaseId } = context.params;
@@ -53,34 +50,15 @@ export function checkout( context, next ) {
 
 	context.store.dispatch( setSection( { name: 'checkout' }, { hasSidebar: false } ) );
 
+	// NOTE: `context.query.code` is deprecated in favor of `context.query.coupon`.
 	const couponCode = context.query.coupon || context.query.code || getRememberedCoupon();
 
-	if ( shouldShowCompositeCheckout() ) {
-		context.primary = (
-			<CartData>
-				<CompositeCheckout
-					siteSlug={ selectedSite?.slug }
-					siteId={ selectedSite?.ID }
-					product={ product }
-					purchaseId={ purchaseId }
-					couponCode={ couponCode }
-					redirectTo={ context.query.redirect_to }
-					feature={ feature }
-				/>
-			</CartData>
-		);
-		next();
-		return;
-	}
-
 	context.primary = (
-		<CheckoutContainer
+		<CheckoutSystemDecider
 			product={ product }
 			purchaseId={ purchaseId }
 			selectedFeature={ feature }
-			// NOTE: `context.query.code` is deprecated in favor of `context.query.coupon`.
-			couponCode={ context.query.coupon || context.query.code || getRememberedCoupon() }
-			// Are we being redirected from the signup flow?
+			couponCode={ couponCode }
 			isComingFromSignup={ !! context.query.signup }
 			plan={ plan }
 			selectedSite={ selectedSite }
@@ -218,17 +196,4 @@ export function redirectToSupportSession( context ) {
 		page.redirect( `/checkout/offer-support-session/${ receiptId }/${ site }` );
 	}
 	page.redirect( `/checkout/offer-support-session/${ site }` );
-}
-
-function shouldShowCompositeCheckout() {
-	if ( ! config.isEnabled( 'composite-checkout-wpcom' ) ) {
-		return false;
-	}
-	// TODO: if a domain product is in the cart, return false
-	// TODO: if a non-wpcom product is in the cart, return false
-	// TODO: if the language is not en-us, return false
-	// TODO: if the geolocation is not in the US, return false
-	if ( abtest( 'showCompositeCheckout' ) === 'composite' ) {
-		return true;
-	}
 }
